moved models to osf_models, added MetaSchema model
